requirements.txt
confluent-kafka==2.6.1
pyspark==3.5.3
faker==30.3.0
mysql-connector-python==9.1.0
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
python-dotenv==1.0.1
SQLAlchemy==2.0.36
PyMySQL==1.1.1
requests==2.32.3
